Removed commented out code
